Update @github/copilot to 1.0.57-2 (#1517)
* Update @github/copilot to 1.0.57-2

- Updated nodejs and test harness dependencies
- Re-ran code generators
- Formatted generated code

* Synthesize shared anyOf discriminator aliases and exclude hand-written collisions

Follows up on #1515. Adds two pieces of generator hardening on top of the
conditional alias emit that #1515 introduced.

1. `collectGoSharedAnyOfDiscriminatorAliasNames`: detect shared anyOf
   unions with a string-const discriminator property and synthesize the
   discriminator enum name + per-variant const names so the public
   `copilot` alias file re-exports them alongside the union and variant
   structs. Without this, schema-shared discriminated unions like
   `Attachment` (in the bumped schema) would expose `copilot.Attachment`
   and its variant structs but not `copilot.AttachmentType` /
   `copilot.AttachmentTypeFile` etc.

2. `collectHandWrittenGoPublicNames`: scan hand-written `go/*.go` files
   for exported type/const names and exclude them from the alias file so
   schema-shared definitions that happen to share a name with a hand-written
   declaration (e.g. `ContextTier`) don't double-declare in the
   `copilot` package and break `go build`. Note from #1515 explicitly
   flagged `ContextTier` as out-of-scope; this handles it.

On the current 1.0.56 schema the only behavioural change is that the
public `copilot` package now also re-exports the
`UserToolSessionApprovalKind` discriminator enum and its 8 constants
(which the `rpc` package was already exporting). On the bumped 1.0.57+
schema this lets the generator succeed (no more
`assertNoGoRpcSessionEventConflicts` failure) and produces the expected
`Attachment*` alias surface while letting hand-written `ContextTier`
win.


* Fix generated type collisions after Copilot update

Remove handwritten ContextTier definitions now owned by generated schema types, update Attachment type references, and keep generated outputs in sync.


* Fix Rust import generation and Python test import ordering

- Rust generator: merge crate::types imports in api_types.rs so the source matches what rustfmt produces on Linux (joining adjacent use crate::types::... lines). This unbreaks the Codegen Check on CI.
- Python: sort the import block in test_event_forward_compatibility.py to satisfy ruff I001.


* Fix Go session event union decoding

Register shared API union decoders when generating Go session event payload unmarshalling so event data can decode attachments and approvals referenced from the API schema.


---------

+++ go/internal/e2e/rpc_event_log_e2e_test.go
@@ -29,7 +29,7 @@ func TestRpcEventLogE2E(t *testing.T) {
 		waitForRPCCondition(t, rpcEventLogTimeout, "persisted session.plan_changed event", func() (bool, error) {
 			var err error
 			read, err = session.RPC.EventLog.Read(t.Context(), &rpc.EventLogReadRequest{
-				Max:    rpcPtr(int32(100)),
+				Max:    rpcPtr(int64(100)),
 				WaitMs: rpcPtr(int32(0)),
 			})
 			if err != nil {
@@ -67,7 +67,7 @@ func TestRpcEventLogE2E(t *testing.T) {
 			}
 			read, err = session.RPC.EventLog.Read(t.Context(), &rpc.EventLogReadRequest{
 				Cursor: &tail.Cursor,
-				Max:    rpcPtr(int32(10)),
+				Max:    rpcPtr(int64(10)),
 				WaitMs: rpcPtr(int32(0)),
 			})
 			return err == nil && read.CursorStatus == rpc.EventsCursorStatusOk && len(read.Events) == 0, err
@@ -131,7 +131,7 @@ func TestRpcEventLogE2E(t *testing.T) {
 			go func() {
 				result, err := session.RPC.EventLog.Read(t.Context(), &rpc.EventLogReadRequest{
 					Cursor: &tail.Cursor,
-					Max:    rpcPtr(int32(10)),
+					Max:    rpcPtr(int64(10)),
 					WaitMs: rpcPtr(int32(5000)),
 					Types:  &rpc.EventLogTypes{StringArray: []string{string(copilot.SessionEventTypeSessionTitleChanged)}},
 				})
